fix bug

+++ lib/.config/rofi/powermenu/powermenu.sh
@@ -34,11 +34,11 @@ uptime=$(uptime -p | sed -e 's/up //g')
 rofi_command="rofi -x11 -theme $dir/full_rounded"
 
 # Options
-shutdown=""
-reboot=""
-lock=""
-suspend=""
-logout=""
+shutdown=""
+reboot=""
+lock=""
+suspend="󰙧"
+logout=""
 
 # Confirmation
 confirm_exit() {
@@ -57,31 +57,48 @@ msg() {
 # Variable passed to rofi
 options="$shutdown\n$reboot\n$lock\n$suspend\n$logout"
 
-chosen="$(echo -e "$options" | $rofi_command -p "Uptime: $uptime" -dmenu -selected-row 2)"
+chosen="$(echo -e "$options" | $rofi_command -p "运行时间: $uptime" -dmenu -selected-row 2)"
 case $chosen in
     $shutdown)
-			systemctl poweroff
+        ans=$(confirm_exit &)
+        if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+            systemctl poweroff
+        elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+            exit 0
+        else
+            msg
+        fi
         ;;
     $reboot)
-			systemctl reboot
+        ans=$(confirm_exit &)
+        if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+            systemctl reboot
+        elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+            exit 0
+        else
+            msg
+        fi
         ;;
     $lock)
-			betterlockscreen -l
+        # 使用 swaylock 锁屏
+        swaylock
         ;;
     $suspend)
-			mpc -q pause
-			amixer set Master mute
-			systemctl suspend
+        # 暂停音频并挂起
+        mpc -q pause 2>/dev/null
+        amixer set Master mute 2>/dev/null
+        # 多种挂起尝试
+        systemctl suspend || loginctl suspend || echo "挂起功能不可用"
         ;;
     $logout)
-			if [[ "$DESKTOP_SESSION" == "Openbox" ]]; then
-				openbox --exit
-			elif [[ "$DESKTOP_SESSION" == "bspwm" ]]; then
-				bspc quit
-			elif [[ "$DESKTOP_SESSION" == "i3" ]]; then
-				i3-msg exit
-			elif [[ "$DESKTOP_SESSION" == "xfce" ]]; then
-				killall xfce4-session
-			fi
+        ans=$(confirm_exit &)
+        if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+            # Hyprland 注销方式
+            hyprctl dispatch exit
+        elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+            exit 0
+        else
+            msg
+        fi
         ;;
 esac
